+++ com.unity.formats.alembic/Runtime/Scripts/Misc/Utils.cs
@@ -34,5 +34,13 @@ namespace UnityEngine.Formats.Alembic.Importer
             var ret = go.GetComponent<T>();
             return ret != null ? ret : go.AddComponent<T>();
         }
+      
+        public static ulong CombineHash(this ulong h1, ulong h2)
+        {
+            unchecked
+            {
+                return h1 ^ h2 + 0x9e3779b9 + (h1 << 6) + (h1 >> 2); // Similar to c++ boost::hash_combine
+            }
+        }
     }
 }