+++ com.unity.formats.alembic/CHANGELOG.md
@@ -4,9 +4,12 @@
 - New option to automatically add the AlembicCurveRendering components for basic preview of the curves in the Scene.
 ### Changed
 - Minimum Unity version is 2019.4
+- Added a custom dependency on the "CurrentRenderPipeline" setting and assigning the correct SRP material by default.
+
 ### Fixed
 - Fixed a bug that caused the Alembic exporter to fail if GameObjects were being deleted during the recording session.
 - Fixed a memory leak that occurred when using varying topology meshes.
+- Fixed a bug that caused the exported mesh normals to be unreadable in DCCs.
 
 ## [2.2.0-exp.2] - 2021-01-21
 ### Added